Create a Dockerfile for Apache ActiveMQ Artemis
The container is based on the OpenJDK 11 image, because Artemis needs
a JRE to be executed. At first, the currently latest version of
Artemis (2.20.0) is downloaded from the official download page, and
the files signature gets checked against the available GPG signature.
Next, the downloaded archive gets extracted into `/usr/local/artemis`,
even though this is probably not the best location following
filesystem conventions, it is right next to the openjdks folder.
After the successful installation of Artemis a new Broker gets created
in `/var/local/lib/artemis-$NAME`. The Brokers `etc` folder gets set
to `/usr/local/etc/artemis-broker` to follow filesystem conventions
more closely. The container sets the environment variables
`ARTEMIS_HOME` and `ARTEMIS_INSTANCE` automatically to the correct
locations. Various other environment variables get used to control the
settings for the created Broker instance, this means that it may not
be necessary to bind the `etc` folder and customize the config files
for simple use cases. The `artemis run` command of the Broker gets set
as cmd and will therefor be executed for a new container as soon as it
starts.

Signed-off-by: Niklas Wimmer <nwimmer@student.tgm.ac.at>